+++ 1.21.6/Common/src/main/java/fuzs/puzzleslib/api/event/v1/server/SyncDataPackContentsCallback.java
@@ -10,7 +10,7 @@ public interface SyncDataPackContentsCallback {
     /**
      * Fires before data pack contents (mainly tags and recipes) are sent to a player. This is caused either by a new
      * player connecting to the server, or when data pack contents have been reloaded, e.g. from running
-     * <code>/reload</code>. In either case the server resource manager is already up-to-date.
+     * <code>/reload</code>. In either case the server resource manager is already up to date.
      *
      * @param serverPlayer the player data pack contents are to be synced with
      * @param joined       are contents being sent because the player just joined, otherwise contents are being sent to