+++ 1.21.6/Fabric/src/main/java/fuzs/puzzleslib/fabric/mixin/client/ClientLevelFabricMixin.java
@@ -93,7 +93,7 @@ abstract class ClientLevelFabricMixin extends Level {
     public void playSeededSound$0(Args args, @Cancellable CallbackInfo callback) {
         Preconditions.checkArgument(args.get(3) instanceof SoundEvent, "sound event is wrong type");
         EventResult eventResult = FabricEventImplHelper.onPlaySound((MutableValue<Holder<SoundEvent>> soundEvent, MutableValue<SoundSource> soundSource, MutableFloat soundVolume, MutableFloat soundPitch) -> {
-                    return FabricLevelEvents.PLAY_LEVEL_SOUND_AT_POSITION.invoker()
+                    return FabricLevelEvents.PLAY_SOUND_AT_POSITION.invoker()
                             .onPlaySoundAtPosition(this,
                                     new Vec3(args.get(0), args.get(1), args.get(2)),
                                     soundEvent,
@@ -120,7 +120,7 @@ abstract class ClientLevelFabricMixin extends Level {
     public void playSeededSound$1(Args args, @Cancellable CallbackInfo callback) {
         Preconditions.checkArgument(args.get(0) instanceof SoundEvent, "sound event is wrong type");
         EventResult eventResult = FabricEventImplHelper.onPlaySound((MutableValue<Holder<SoundEvent>> soundEvent, MutableValue<SoundSource> soundSource, MutableFloat soundVolume, MutableFloat soundPitch) -> {
-                    return FabricLevelEvents.PLAY_LEVEL_SOUND_AT_ENTITY.invoker()
+                    return FabricLevelEvents.PLAY_SOUND_AT_ENTITY.invoker()
                             .onPlaySoundAtEntity(this, args.get(4), soundEvent, soundSource, soundVolume, soundPitch);
                 },
                 args,